Give utils functions their own module
Keep ToucanClient simple.
