+++ tests/__init__.py
@@ -1,0 +1,8 @@
+SMALL_APP_NAME = 'my-small-app'
+BASE_ROUTE = 'fake.route/my-small-app'
+METHODS = {
+    'get': 1,
+    'post': 2,
+    'put': 3,
+    'delete': 4,
+}
